Update handling of transparent reconnections
The previous version of the RFC stated that a transparent reconnection
would generate a single Reconnect event (not preceded by Reconnecting).
This caused two issues: it is trickier to implement, and it makes
reconnection events inconsistent. Because of these issues and because
the Reconnect event on a transparent reconnection was purely for
informational purposes (users couldn't act on it in any meaningful way),
it was decided to remove that feature from the RFC.